fix mask to image as array of uint8

+++ slaid/commons/base.py
@@ -122,7 +122,7 @@ class Mask:
         array = self.array[::downsample, ::downsample]
         if threshold:
             array = apply_threshold(array, threshold)
-        return PIL.Image.fromarray(255 * array, 'L')
+        return PIL.Image.fromarray((255 * array).astype('uint8'), 'L')
 
     def to_polygons(self,
                     threshold: float = None,
